misc: Move some functions to util, types to types
